# testing-environment-backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/projects": {
      "post": {
        "operationId": "ProjectsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Projects"
        ]
      },
      "get": {
        "operationId": "ProjectsController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedProjectsResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Projects"
        ]
      }
    },
    "/projects/{id}": {
      "get": {
        "operationId": "ProjectsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Projects"
        ]
      },
      "patch": {
        "operationId": "ProjectsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Projects"
        ]
      },
      "delete": {
        "operationId": "ProjectsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Projects"
        ]
      }
    },
...
    "/team/members": {
      "get": {
        "operationId": "TeamController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TeamMemberResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Team"
        ]
      }
    },
    "/team/members/{memberId}": {
      "patch": {
        "operationId": "TeamController_updateRole",
        "parameters": [
          {
            "name": "memberId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMemberRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamMemberResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Team"
        ]
      },
      "delete": {
        "operationId": "TeamController_remove",
        "parameters": [
          {
            "name": "memberId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Team"
        ]
      }
    },
...
    "/projects/{projectId}/test-runs": {
      "post": {
        "operationId": "TestRunsController_create",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTestRunDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestRunResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Test runs"
        ]
      },
      "get": {
        "operationId": "TestRunsController_list",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedTestRunsResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Test runs"
        ]
      }
    },
    "/projects/{projectId}/test-runs/{runId}": {
      "get": {
        "operationId": "TestRunsController_find",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "runId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestRunResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Test runs"
        ]
      }
    },
    "/projects/{projectId}/test-runs/{runId}/events": {
      "get": {
        "operationId": "TestRunsController_events",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "runId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "afterSequence",
            "required": false,
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TestRunEventResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Test runs"
        ]
      }
    },
    "/projects/{projectId}/test-runs/{runId}/cancel": {
      "post": {
        "operationId": "TestRunsController_cancel",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "runId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelTestRunDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestRunResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Test runs"
        ]
      }
    },
...
  },
  "info": {
    "title": "Backend Test Runner API",
    "description": "MVP backend control plane and local runner engine.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "Object": {
        "type": "object",
        "properties": {}
      },
...
      "PaginationMetaDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          },
          "total": {
            "type": "number"
          },
          "totalPages": {
            "type": "number"
          }
        },
        "required": [
          "page",
          "limit",
          "total",
          "totalPages"
        ]
      },
...
      "CreateProjectDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "baseUrl": {
            "type": "string"
          },
          "mainServiceName": {
            "type": "string"
          },
          "healthcheckPath": {
            "type": "string",
            "default": "/health"
          },
          "healthcheckExpectedStatus": {
            "type": "number",
            "default": 200
          },
          "healthcheckTimeoutSeconds": {
            "type": "number",
            "default": 60
          }
        },
        "required": [
          "name",
          "baseUrl",
          "mainServiceName",
          "healthcheckPath",
          "healthcheckExpectedStatus",
          "healthcheckTimeoutSeconds"
        ]
      },
      "ProjectResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "companyId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "baseUrl": {
            "type": "string"
          },
          "mainServiceName": {
            "type": "string"
          },
          "healthcheckPath": {
            "type": "string"
          },
          "healthcheckExpectedStatus": {
            "type": "number"
          },
          "healthcheckTimeoutSeconds": {
            "type": "number"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "companyId",
          "name",
          "baseUrl",
          "mainServiceName",
          "healthcheckPath",
          "healthcheckExpectedStatus",
          "healthcheckTimeoutSeconds",
          "createdAt",
          "updatedAt"
        ]
      },
      "PaginatedProjectsResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProjectResponseDto"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/PaginationMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "UpdateProjectDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "baseUrl": {
            "type": "string"
          },
          "mainServiceName": {
            "type": "string"
          },
          "healthcheckPath": {
            "type": "string",
            "default": "/health"
          },
          "healthcheckExpectedStatus": {
            "type": "number",
            "default": 200
          },
          "healthcheckTimeoutSeconds": {
            "type": "number",
            "default": 60
          }
        }
      },
...
      "TeamMemberUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "email",
          "firstName",
          "lastName",
          "createdAt",
          "updatedAt"
        ]
      },
      "TeamMemberResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "companyId": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "OWNER",
              "ADMIN",
              "DEVELOPER",
              "VIEWER"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "REMOVED"
            ]
          },
          "removedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "user": {
            "$ref": "#/components/schemas/TeamMemberUserDto"
          }
        },
        "required": [
          "id",
          "companyId",
          "userId",
          "role",
          "status",
          "createdAt",
          "updatedAt",
          "user"
        ]
      },
      "UpdateMemberRoleDto": {
        "type": "object",
        "properties": {
          "role": {
            "type": "string",
            "enum": [
              "OWNER",
              "ADMIN",
              "DEVELOPER",
              "VIEWER"
            ]
          }
        },
        "required": [
          "role"
        ]
      },
...
      "CreateTestRunDto": {
        "type": "object",
        "properties": {
          "environmentConfigRevisionId": {
            "type": "string",
            "description": "Explicit environment revision to run against. Defaults to latest published revision."
          }
        }
      },
      "TestRunResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "projectId": {
            "type": "string"
          },
          "environmentConfigRevisionId": {
            "type": "string",
            "nullable": true
          },
          "runnerVersion": {
            "type": "string"
          },
          "reportSchemaVersion": {
            "type": "number"
          },
          "status": {
            "type": "string",
            "enum": [
              "CREATED",
              "QUEUED",
              "CLAIMED",
              "PREPARING_WORKSPACE",
              "VALIDATING_ENVIRONMENT",
              "PULLING_IMAGES",
              "STARTING_ENVIRONMENT",
              "WAITING_FOR_HEALTHCHECK",
              "EXECUTING_TESTS",
              "COLLECTING_ARTIFACTS",
              "CLEANING_UP",
              "PASSED",
              "TEST_FAILED",
              "INFRA_FAILED",
              "TIMED_OUT",
              "CANCEL_REQUESTED",
              "CANCELLED"
            ]
          },
          "statusReason": {
            "type": "string",
            "description": "Human-readable terminal or cancellation reason for the current run state.",
            "nullable": true
          },
          "failureCategory": {
            "type": "string",
            "enum": [
              "TEST_ASSERTION",
              "ENVIRONMENT_VALIDATION",
              "IMAGE_PULL",
              "CONTAINER_START",
              "HEALTHCHECK",
              "NETWORK",
              "TIMEOUT",
              "CANCELLED",
              "INTERNAL"
            ],
            "nullable": true
          },
          "currentPhase": {
            "type": "string",
            "nullable": true
          },
          "phaseTimestamps": {
            "type": "object"
          },
          "queueJobId": {
            "type": "string",
            "nullable": true
          },
          "queuedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "enqueuedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "claimedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "cancellationRequestedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "cancelRequestedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "cancelRequestedBy": {
            "type": "string",
            "nullable": true
          },
          "cancellationReason": {
            "type": "string",
            "nullable": true
          },
          "runnerId": {
            "type": "string",
            "nullable": true
          },
          "leaseAcquiredAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "leaseExpiresAt": {
            "type": "string",
            "description": "Current execution lease expiration; stale leases are failed by recovery.",
            "format": "date-time",
            "nullable": true
          },
          "heartbeatAt": {
            "type": "string",
            "description": "Last persisted worker heartbeat for the execution lease owner.",
            "format": "date-time",
            "nullable": true
          },
          "attempt": {
            "type": "number",
            "description": "Number of worker claims for this immutable run. Jobs are fail-fast by default."
          },
          "cleanupError": {
            "type": "string",
            "nullable": true
          },
          "startedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "finishedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "totalTests": {
            "type": "number"
          },
          "passedTests": {
            "type": "number"
          },
          "failedTests": {
            "type": "number"
          },
          "durationMs": {
            "type": "number",
            "nullable": true
          },
          "errorMessage": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "projectId",
          "runnerVersion",
          "reportSchemaVersion",
          "status",
          "attempt",
          "totalTests",
          "passedTests",
          "failedTests"
        ]
      },
      "PaginatedTestRunsResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TestRunResponseDto"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/PaginationMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "TestRunEventResponseDto": {
        "type": "object",
        "properties": {
          "runId": {
            "type": "string"
          },
          "sequence": {
            "type": "number"
          },
          "type": {
            "type": "string"
          },
          "timestamp": {
            "type": "string"
          },
          "payload": {
            "type": "object"
          }
        },
        "required": [
          "runId",
          "sequence",
          "type",
          "timestamp",
          "payload"
        ]
      },
      "CancelTestRunDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "maxLength": 500
          }
        }
      }
    }
  }
}